RealSense Person Turtlebot Follower
This package contains a nodelet that connects the Intel(R) RealSense(TM)
Person middleware with the Turtlebot Follower.

The RealSensePersonCoM nodelet subscribes to the detection_data
topic produced by the realsense_person middleware. It then extracts
the Center of Mass of the first person reported and uses this to
publish a geometry_msgs/Point goal message.

This nodelet is designed to be a swap-in replacement for the
centroid goal-generating nodelet in the turtlebot_follower package.
The included launch file launches the realsense_person middleware
and the turtlebot_follower application, and maps the goal output
appropriately.