Add gesture detection to stop robot.
Adds new functionality to the Turtlebot Follower, allowing you to stop the Turtlebot from
following you by performing a simple gesture.

Utilizes the new person-tracking functionality of the Realsense Person middleware to track
the pose of the person being followed. If the person raises their either hand above their
center of mass, the robot is stopped and goal messages are not published. Once the person
lowers their hand, following is resumed.

This greatly simplifies the process of stopping the turtlebot from following you, as you can
now simply raise your hand in the air and walk out of the frame of view.